+++ packages/cloud-cognitive/src/components/AddSelect/AddSelectSort.js
@@ -10,11 +10,11 @@ import PropTypes from 'prop-types';
 import { OverflowMenu, OverflowMenuItem } from 'carbon-components-react';
 import { ArrowsVertical32, ArrowUp16, ArrowDown16 } from '@carbon/icons-react';
 import { pkg } from '../../settings';
-const componentName = 'AddSelect';
+
+const blockClass = `${pkg.prefix}--add-select-sort`;
+const componentName = 'AddSelectSort';
 
 export let AddSelectSort = ({ setSortAttribute, setSortDirection, sortBy }) => {
-  const blockClass = `${pkg.prefix}--add-select-sort`;
-
   const sortByOpts = sortBy
     ? sortBy.reduce((acc, cur) => {
         const opts = [
@@ -54,9 +54,14 @@ export let AddSelectSort = ({ setSortAttribute, setSortDirection, sortBy }) => {
   return (
     <div className={blockClass}>
       {sortByOpts.length > 0 && (
-        <OverflowMenu renderIcon={ArrowsVertical32} flipped>
+        <OverflowMenu
+          className={`${blockClass}_overflow`}
+          renderIcon={ArrowsVertical32}
+          flipped
+        >
           {sortByOpts.map((opt) => (
             <OverflowMenuItem
+              className={`${blockClass}_overflow-item`}
               key={opt.id}
               itemText={opt.itemText}
               onClick={() => sortHandler(opt)}